pyproject.toml
[project]
name = "openai-agents-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "display>=1.0.0",
    "openai-agents[viz]>=0.0.16",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
